Refactor logger to accept variable arguments
Changed logging functions in the codebase to make use of a variable argument's template. This allows better string formatting and reduces the need for explicit string concatenation in logger statements, leading to cleaner and more readable code.

+++ editor/ui/net/UIManager.h
@@ -33,7 +33,7 @@ namespace moonshine {
         std::mutex uiMap;
 
         float since_last_updated = 0;
-        float rate_limit = 50;
+        int rate_limit = 50;
 
     public:
